btrfs-progs: move volumes.c to kernel-shared/
