+++ btrfs-corrupt-block.c
@@ -25,7 +25,7 @@
 
 #include "kerncompat.h"
 #include "kernel-shared/ctree.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "kernel-shared/transaction.h"
