+++ btrfs-select-super.c
@@ -23,7 +23,7 @@
 #include <sys/stat.h>
 #include "kerncompat.h"
 #include "kernel-shared/ctree.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "kernel-shared/transaction.h"
