+++ btrfstune.c
@@ -31,7 +31,7 @@
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/transaction.h"
 #include "common/utils.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "common/help.h"
 #include "common/box.h"
