+++ cmds/filesystem.c
@@ -34,7 +34,7 @@
 #include "kerncompat.h"
 #include "kernel-shared/ctree.h"
 #include "common/utils.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "cmds/commands.h"
 #include "cmds/filesystem-usage.h"
 #include "kernel-lib/list_sort.h"
