+++ cmds/inspect-dump-tree.c
@@ -29,7 +29,7 @@
 #include "kernel-shared/disk-io.h"
 #include "kernel-shared/print-tree.h"
 #include "kernel-shared/transaction.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "cmds/commands.h"
 #include "common/utils.h"
 #include "common/help.h"
