+++ cmds/rescue-chunk-recover.c
@@ -34,7 +34,7 @@
 #include "kernel-shared/ctree.h"
 #include "common/extent-cache.h"
 #include "kernel-shared/disk-io.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/transaction.h"
 #include "crypto/crc32c.h"
 #include "common/utils.h"
