+++ cmds/rescue-super-recover.c
@@ -32,7 +32,7 @@
 #include "kernel-lib/list.h"
 #include "common/utils.h"
 #include "crypto/crc32c.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "cmds/commands.h"
 #include "cmds/rescue.h"
