+++ cmds/restore.c
@@ -42,7 +42,7 @@
 #include "kernel-shared/print-tree.h"
 #include "kernel-shared/transaction.h"
 #include "kernel-lib/list.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "common/utils.h"
 #include "cmds/commands.h"
 #include "common/help.h"
