+++ common/device-scan.c
@@ -33,7 +33,7 @@
 #include "common/utils.h"
 #include "common/defs.h"
 #include "kernel-shared/ctree.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/disk-io.h"
 #include "ioctl.h"
