+++ convert/main.c
@@ -93,7 +93,7 @@
 
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/transaction.h"
 #include "common/utils.h"
 #include "common/task-utils.h"
