+++ kernel-shared/ctree.c
@@ -24,7 +24,7 @@
 #include "common/messages.h"
 #include "common/utils.h"
 #include "kernel-lib/sizes.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 
 static int split_node(struct btrfs_trans_handle *trans, struct btrfs_root
 		      *root, struct btrfs_path *path, int level);
