+++ kernel-shared/disk-io.c
@@ -27,7 +27,7 @@
 #include "kernel-lib/radix-tree.h"
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "kernel-shared/transaction.h"
 #include "crypto/crc32c.h"
 #include "common/utils.h"
