+++ mkfs/common.c
@@ -21,7 +21,7 @@
 #include <limits.h>
 #include "kernel-shared/ctree.h"
 #include "kernel-shared/disk-io.h"
-#include "volumes.h"
+#include "kernel-shared/volumes.h"
 #include "common/utils.h"
 #include "common/path-utils.h"
 #include "common/device-utils.h"
